Move proxy-select-field to a common area for reuse by resource and package
